Experiment with strongly typed labels
Overall it appears doable but somewhat complex for marginal gains.
The strong typing makes it harder to make a mistake at the recording
callsite, but at the cost of making construction and invocation more
verbose, and considerably more machinery in the runtime library. The
listener and SDK also have to opt into a new ABI or there is overall perf
loss if the strong type ever converts back to string[].To support AOT
compile/link we'd probably need to add a code generator and for the SDK
to preserve high perf it too would need to do some non-trivial work
to avoid needing to convert to string[].